chore: move user list to the correct component folder
